+++ src/app/globals.css
@@ -5,27 +5,22 @@
 @media (prefers-color-scheme: dark) {
   :root {
     --background-color: 10 10 10;
-    --background-color-disabled: 10 10 10 / 0.6;
     --foreground-color: 237 237 237;
-    --foreground-color-disabled: 237 237 237 / 0.6;
   }
 }
 
 @layer base {
   :root {
-    /* Layout */
     --background-color: 255 255 255;
-    --background-color-disabled: 255 255 255 / 0.6;
     --foreground-color: 23 23 23;
-    --foreground-color-disabled: 23 23 23 / 0.6;
 
-    /* Bg Colors */
+    /* Color Palette */
     --primary-color: 15 23 42; /* bg-slate-900 */
     --secondary-color: 30 41 59; /* bg-slate-800 */
     --tertiary-color: 51 65 85; /* bg-slate-700 */
     --accent-color: 71 85 105; /* bg-slate-600 */
 
-    /* Btn Variants */
+    /* Btn Variant Colors */
     --btn-ghost-bg: 23 37 84 / 0.2; /* bg-cyan-500 */
     --btn-primary-bg: 21 128 61; /* bg-green-700 */
     --btn-warning-bg: 185 28 28; /* bg-red-700 */
