--- src/tw-styled/theme/index.ts
@@ -1,3 +1,0 @@
-export * from "./theme";
-export * from "./ThemeProvider";
-export * from "./useTheme";
