--- src/tw-styled/theme/useTheme.ts
@@ -1,3 +1,0 @@
-export const useTheme = (props: {}) => {
-  return {};
-};
