+++ src/tw-styled/theme/useThemeCtx.ts
@@ -1,0 +1,3 @@
+export const useThemeCtx = (props: {}) => {
+  return {};
+};
