+++ src/tw-styled/tools.ts
@@ -1,4 +1,3 @@
 export * from "./utils/class-merge";
 export * from "./hooks/useIcons";
 export * from "./style-resolver/hooks/useStyleResolver";
-export * from "./theme";
