+++ src/tw-styled/tw.ts
@@ -1,0 +1,2 @@
+export * from "./tailwind/config";
+export * from "./tailwind/useRgb";
