+++ tailwind.config.ts
@@ -1,24 +1,20 @@
-import type { Config } from "tailwindcss";
 import plugin from "tailwindcss/plugin";
-import { components, utilities } from "@/tw-styled/tailwind/config";
-import { themeCtx } from "@/tw-styled/theme";
-
-const useRgb = (variable: string, opacity?: number) =>
-  opacity ? `rgb(var(${variable}) / ${opacity})` : `rgb(var(${variable}))`;
+import type { Config } from "tailwindcss";
+import { components, useRgb, utilities } from "./src/tw-styled/tw";
+import { themeCtx } from "./src/tw-styled/theme/themeCtx";
 
 const config: Config = {
   content: [
     "./src/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/**/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/tw-styled/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/tw-styled/stories/**/*.{js,ts,jsx,tsx,mdx}"
+    "./src/tw-styled/**/**/*.{js,ts,jsx,tsx,mdx}"
   ],
   theme: {
     extend: {
       colors: {
         background: useRgb("--background-color"),
-        "background-disabled": useRgb("--background-color-disabled"),
-        foreground: useRgb("--foreground-color"),
-        "foreground-disabled": useRgb("--foreground-color-disabled")
+        foreground: useRgb("--foreground-color")
       },
       backgroundColor: {
         primary: useRgb(themeCtx.colors.primary),
